requirements.txt
Flask==3.0.0
mysql-connector-python==8.2.0
weasyprint==66.0
openpyxl==3.1.2
pandas==2.1.3
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.0
